requirements.txt
torch
numpy
torchvision
Pillow
opencv-python
matplotlib
requests
